Rename environment files for better readability
As pointed out in code review by @lencioni.
https://github.com/Galooshi/import-js/pull/302/files/c5c43cf2178c66ae3219a3373373b6198d77b641..7d8dfbff30fe5bb3809fd1c8b78a6dce1705bc83#r70171899
