+++ lib/Configuration.js
@@ -7,8 +7,8 @@ import semver from 'semver';
 
 import FileUtils from './FileUtils';
 import JsModule from './JsModule';
-import meteor from './environments/meteor';
-import node from './environments/node';
+import meteorEnvironment from './environments/meteorEnvironment';
+import nodeEnvironment from './environments/nodeEnvironment';
 import normalizePath from './normalizePath';
 import requireResolve from './requireResolve';
 import version from './version';
@@ -84,8 +84,8 @@ const KNOWN_CONFIGURATION_OPTIONS = [
 ];
 
 const ENVIRONMENTS = {
-  node,
-  meteor,
+  node: nodeEnvironment,
+  meteor: meteorEnvironment,
 };
 
 function convertRenamedConfiguration(config: Object): Object {
